requirements.txt
gspread==0.1.0
pyserial==2.7
redis==2.9.1
